requirements.txt
Flask==1.1.2
google-oauth==1.0.1
google-cloud-dialogflow==2.0.0
Flask-Caching==1.9.0
pyTelegramBotAPI==3.7.4
